… opentype/tables/glyphs_test.go
… opentype/tables/ot_layout.go
… opentype/tables/ot_layout_test.go
… opentype/tables/ot_properties.go
… shaping/shaping_test.go