The general approach used in the Opentype parser is to aggressively decode the binary format into Go types, which are clearer and simpler to use that the intricate, offset-based Opentype format.

However, the Opentype format is intricate precisely because it uses clever conventions to very efficiently compact glyph data : decoding this data yields an in-memory representation that is larger.

Recent issues regarding memory usage indicate that we should aim at a better balance :  this PR reduces memory usage for the `glyf` and `GPOS` tables, leading to significant savings for large fonts.
For instance, the [font](https://github.com/satbyy/go-noto-universal/releases/download/v5.3/GoNotoCurrent.ttf) mentionned in this Fyne [issue](https://github.com/fyne-io/fyne/issues/3499), 14Mb on disk, was taking roughly 70Mb in memory, now reduced to 24Mb !

The speed during shaping is affected in the following ways : 

- slightly slower shaping for a few  `GPOS` operations 
- shaping otherwise not affected for non variable font (which is always the case for now)
- slower glyph rendering (via the `face.GlyphData` method)

... but the memory savings seem to be worth it.