@whereswaldon 2023-04-10
I've taken the liberty of adding a benchmark commit prior to this changeset and measuring the impact of this branch. Here are my results:

```
goos: linux
goarch: amd64
pkg: github.com/go-text/typesetting/shaping
cpu: Intel(R) Core(TM) i7-6700K CPU @ 4.00GHz
                                              │ main-branch.txt │        lowermem-branch.txt         │
                                              │     sec/op      │   sec/op     vs base               │
Shaping/10runes-arabic-0fontCache-8                 155.7µ ± 1%   160.8µ ± 1%   +3.28% (p=0.001 n=7)
Shaping/100runes-arabic-0fontCache-8                458.4µ ± 1%   468.2µ ± 1%   +2.15% (p=0.001 n=7)
Shaping/1000runes-arabic-0fontCache-8               3.369m ± 3%   3.461m ± 2%   +2.72% (p=0.011 n=7)
Shaping/10runes-latin-0fontCache-8                  2.671µ ± 1%   2.833µ ± 1%   +6.07% (p=0.001 n=7)
Shaping/100runes-latin-0fontCache-8                 17.16µ ± 1%   18.28µ ± 1%   +6.49% (p=0.001 n=7)
Shaping/1000runes-latin-0fontCache-8                160.9µ ± 1%   171.1µ ± 0%   +6.33% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-arabic-8         23.58µ ± 1%   30.07µ ± 2%  +27.54% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-arabic-8        224.9µ ± 1%   290.8µ ± 1%  +29.34% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-arabic-8       2.304m ± 1%   2.978m ± 1%  +29.23% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-latin-8          10.05µ ± 2%   12.63µ ± 8%  +25.68% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-latin-8         101.4µ ± 2%   128.0µ ± 1%  +26.31% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-latin-8        1.007m ± 1%   1.280m ± 1%  +27.03% (p=0.001 n=7)
geomean                                             109.4µ        121.9µ       +11.44%

                                              │ main-branch.txt │         lowermem-branch.txt          │
                                              │      B/op       │     B/op      vs base                │
Shaping/10runes-arabic-0fontCache-8                68.92Ki ± 0%   68.92Ki ± 0%       ~ (p=0.266 n=7)
Shaping/100runes-arabic-0fontCache-8               136.4Ki ± 0%   136.4Ki ± 0%       ~ (p=0.674 n=7)
Shaping/1000runes-arabic-0fontCache-8              858.1Ki ± 0%   858.1Ki ± 0%  +0.00% (p=0.017 n=7)
Shaping/10runes-latin-0fontCache-8                 2.234Ki ± 0%   2.234Ki ± 0%       ~ (p=1.000 n=7) ¹
Shaping/100runes-latin-0fontCache-8                7.984Ki ± 0%   7.984Ki ± 0%       ~ (p=1.000 n=7) ¹
Shaping/1000runes-latin-0fontCache-8               65.64Ki ± 0%   65.64Ki ± 0%       ~ (p=0.085 n=7)
ShapingGlyphInfoExtraction/10runes-arabic-8        54.85Ki ± 0%   57.71Ki ± 0%  +5.21% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-arabic-8       499.1Ki ± 0%   524.4Ki ± 0%  +5.08% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-arabic-8      4.916Mi ± 0%   5.166Mi ± 0%  +5.08% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-latin-8         19.20Ki ± 0%   20.26Ki ± 0%  +5.54% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-latin-8        196.4Ki ± 0%   207.0Ki ± 0%  +5.40% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-latin-8       1.897Mi ± 0%   2.000Mi ± 0%  +5.44% (p=0.001 n=7)
geomean                                            74.07Ki        75.35Ki       +1.73%
¹ all samples are equal

                                              │ main-branch.txt │         lowermem-branch.txt          │
                                              │    allocs/op    │  allocs/op   vs base                 │
Shaping/10runes-arabic-0fontCache-8                 1.363k ± 0%   1.363k ± 0%        ~ (p=1.000 n=7) ¹
Shaping/100runes-arabic-0fontCache-8                3.339k ± 0%   3.339k ± 0%        ~ (p=1.000 n=7) ¹
Shaping/1000runes-arabic-0fontCache-8               24.57k ± 0%   24.57k ± 0%   +0.01% (p=0.001 n=7)
Shaping/10runes-latin-0fontCache-8                   9.000 ± 0%    9.000 ± 0%        ~ (p=1.000 n=7) ¹
Shaping/100runes-latin-0fontCache-8                  9.000 ± 0%    9.000 ± 0%        ~ (p=1.000 n=7) ¹
Shaping/1000runes-latin-0fontCache-8                 9.000 ± 0%    9.000 ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/10runes-arabic-8          163.0 ± 0%    183.0 ± 0%  +12.27% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-arabic-8        1.625k ± 0%   1.825k ± 0%  +12.31% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-arabic-8       16.81k ± 0%   18.93k ± 0%  +12.59% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-latin-8           105.0 ± 0%    114.0 ± 0%   +8.57% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-latin-8         1.028k ± 0%   1.113k ± 0%   +8.27% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-latin-8        10.25k ± 0%   11.10k ± 0%   +8.30% (p=0.001 n=7)
geomean                                              315.4         325.9        +3.34%
¹ all samples are equal
```

Based on my results, this changeset appears to:

- make shaping latin text about 6% slower
- make shaping arabic text 2-3% slower
- have no impact on memory requirements during the shaping process
- make glyph data extraction 25-30% slower for all languages
- increase glyph data extraction memory requirements by about 5% for all languages
- increase the quantity of allocations for arabic glyph data extraction by about 12%
- increase the quantity of allocations for latin glyph data extraction by about 8%

I can live with the slowdown on shaping, but I worry about the glyph data extraction penalty. 25-30% is quite steep. I also (as I was writing this) realized that we have no benchmark to capture the metric this changeset primarily addresses (font loading memory requirements). I'll try to add one for that to measure the benefit on that side a little more concretely.