@whereswaldon 2023-04-10
Okay, I've pushed a font parsing benchmark as well, though it only loads the three fonts available in the typesetting repo, and therefore may not be really representative of our performance across a wide swathe of fonts. Here are my results from that:

```
goos: linux
goarch: amd64
pkg: github.com/go-text/typesetting/shaping
cpu: Intel(R) Core(TM) i7-6700K CPU @ 4.00GHz
                                        │ main-font-parse.txt │      lowermem-font-parse.txt       │
                                        │       sec/op        │   sec/op     vs base               │
ShapingFontLoad/arabic:amiri_regular-8            3.984m ± 2%   2.122m ± 4%  -46.74% (p=0.001 n=7)
ShapingFontLoad/lating:go_regular-8               399.5µ ± 3%   239.8µ ± 3%  -39.98% (p=0.001 n=7)
ShapingFontLoad/lating:roboto_regular-8          1224.5µ ± 3%   335.1µ ± 3%  -72.64% (p=0.001 n=7)
geomean                                           1.249m        554.4µ       -55.61%

                                        │ main-font-parse.txt │       lowermem-font-parse.txt       │
                                        │        B/op         │     B/op      vs base               │
ShapingFontLoad/arabic:amiri_regular-8           4.678Mi ± 0%   2.854Mi ± 0%  -38.98% (p=0.001 n=7)
ShapingFontLoad/lating:go_regular-8              366.7Ki ± 0%   273.6Ki ± 0%  -25.38% (p=0.001 n=7)
ShapingFontLoad/lating:roboto_regular-8         1857.0Ki ± 0%   392.2Ki ± 0%  -78.88% (p=0.001 n=7)
geomean                                          1.448Mi        679.5Ki       -54.18%

                                        │ main-font-parse.txt │      lowermem-font-parse.txt       │
                                        │      allocs/op      │  allocs/op   vs base               │
ShapingFontLoad/arabic:amiri_regular-8            69.11k ± 0%   30.63k ± 0%  -55.69% (p=0.001 n=7)
ShapingFontLoad/lating:go_regular-8               2.764k ± 0%   2.056k ± 0%  -25.62% (p=0.001 n=7)
ShapingFontLoad/lating:roboto_regular-8          13.503k ± 0%   2.599k ± 0%  -80.75% (p=0.001 n=7)
geomean                                           13.71k        5.470k       -60.12%
```

- Font load speed is significantly faster, up to 72% in one case.
- Font load memory use is down between 25-78% depending upon the face.
- Font load allocation count is down 25-80% as well, depending upon the face.

These results make sense. As I understand it, the changeset primarily just does less unpacking into go types, so the results reflect doing less work.

As we only parse a font once (hopefully), whereas we use it continuously, I do think I'm still concerned about the glyph data extraction slowdown. We cache the results of that operation in Gio, but it's still quite expensive.

@benoitkugler Naturally, this is all a balancing act. Reducing memory use will put pressure on something else. Maybe this is the best option we have. Do you see other options that seem worth exploring in terms of balancing our various resource uses, or is this the first logical step from our previous strategy (nothing in between worth exploring)?