@whereswaldon 2023-04-11
> > @benoitkugler Would you mind posting another PR with just the GPOS changes? We can evaluate the benchmarks from just doing that and see what the tradeoffs are.
> 
> Done ! I've also added another measure of the heap size of the resulting `font.Font`, not taking into account intermediate allocations. This is actually the primary metric I wanted to handle.

I've taken benchmarks before/after on the other branch. I'll share them here to keep all of the discussion in one place:

```
goos: linux
goarch: amd64
pkg: github.com/go-text/typesetting/shaping
cpu: Intel(R) Core(TM) i7-6700K CPU @ 4.00GHz
                                              │ main-no-gpos.txt │          lowermem-gpos.txt          │
                                              │      sec/op      │    sec/op     vs base               │
Shaping/10runes-arabic-0fontCache-8                  155.4µ ± 3%   156.9µ ±  1%        ~ (p=0.383 n=7)
Shaping/100runes-arabic-0fontCache-8                 463.9µ ± 2%   463.5µ ±  1%        ~ (p=1.000 n=7)
Shaping/1000runes-arabic-0fontCache-8                3.391m ± 2%   3.395m ±  1%        ~ (p=0.383 n=7)
Shaping/10runes-latin-0fontCache-8                   2.657µ ± 1%   2.673µ ±  3%        ~ (p=0.119 n=7)
Shaping/100runes-latin-0fontCache-8                  17.02µ ± 2%   17.20µ ±  1%   +1.05% (p=0.012 n=7)
Shaping/1000runes-latin-0fontCache-8                 160.6µ ± 2%   163.4µ ±  1%        ~ (p=0.097 n=7)
ShapingFontLoad/arabic:amiri_regular-8               4.004m ± 2%   2.920m ±  1%  -27.07% (p=0.001 n=7)
ShapingFontLoad/lating:go_regular-8                  403.6µ ± 3%   398.5µ ±  1%        ~ (p=0.053 n=7)
ShapingFontLoad/lating:roboto_regular-8             1226.0µ ± 1%   534.0µ ±  1%  -56.44% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-arabic-8          23.50µ ± 2%   24.76µ ±  2%   +5.34% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-arabic-8         226.3µ ± 3%   237.9µ ±  3%   +5.15% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-arabic-8        2.302m ± 1%   2.438m ±  2%   +5.90% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-latin-8           10.03µ ± 3%   10.86µ ± 11%   +8.33% (p=0.001 n=7)
ShapingGlyphInfoExtraction/100runes-latin-8          101.4µ ± 6%   108.2µ ± 15%   +6.80% (p=0.004 n=7)
ShapingGlyphInfoExtraction/1000runes-latin-8         1.015m ± 6%   1.087m ±  3%   +7.17% (p=0.001 n=7)
geomean                                              28.03µ        27.26µ         -2.75%

                                              │ main-no-gpos.txt │           lowermem-gpos.txt           │
                                              │       B/op       │     B/op      vs base                 │
Shaping/10runes-arabic-0fontCache-8               68.92Ki ± 0%     68.92Ki ± 0%        ~ (p=1.000 n=7)
Shaping/100runes-arabic-0fontCache-8              136.4Ki ± 0%     136.4Ki ± 0%        ~ (p=0.592 n=7)
Shaping/1000runes-arabic-0fontCache-8             858.1Ki ± 0%     858.1Ki ± 0%   +0.00% (p=0.004 n=7)
Shaping/10runes-latin-0fontCache-8                2.234Ki ± 0%     2.234Ki ± 0%        ~ (p=1.000 n=7) ¹
Shaping/100runes-latin-0fontCache-8               7.984Ki ± 0%     7.984Ki ± 0%        ~ (p=1.000 n=7) ¹
Shaping/1000runes-latin-0fontCache-8              65.64Ki ± 0%     65.64Ki ± 0%        ~ (p=0.429 n=7)
ShapingFontLoad/arabic:amiri_regular-8            4.678Mi ± 0%     3.484Mi ± 0%  -25.52% (p=0.001 n=7)
ShapingFontLoad/lating:go_regular-8               366.7Ki ± 0%     366.7Ki ± 0%   +0.01% (p=0.001 n=7)
ShapingFontLoad/lating:roboto_regular-8          1857.0Ki ± 0%     503.7Ki ± 0%  -72.88% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-arabic-8       54.85Ki ± 0%     54.85Ki ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/100runes-arabic-8      499.1Ki ± 0%     499.1Ki ± 0%   +0.00% (p=0.001 n=7)
ShapingGlyphInfoExtraction/1000runes-arabic-8     4.916Mi ± 0%     4.916Mi ± 0%   +0.00% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-latin-8        19.20Ki ± 0%     19.20Ki ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/100runes-latin-8       196.4Ki ± 0%     196.4Ki ± 0%        ~ (p=0.070 n=7)
ShapingGlyphInfoExtraction/1000runes-latin-8      1.897Mi ± 0%     1.897Mi ± 0%   +0.00% (p=0.001 n=7)
geomean                                                        ²                  -3.42%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                              │ main-no-gpos.txt │          lowermem-gpos.txt           │
                                              │    allocs/op     │  allocs/op   vs base                 │
Shaping/10runes-arabic-0fontCache-8                1.363k ± 0%     1.363k ± 0%        ~ (p=1.000 n=7) ¹
Shaping/100runes-arabic-0fontCache-8               3.339k ± 0%     3.339k ± 0%        ~ (p=1.000 n=7) ¹
Shaping/1000runes-arabic-0fontCache-8              24.57k ± 0%     24.57k ± 0%   +0.01% (p=0.001 n=7)
Shaping/10runes-latin-0fontCache-8                  9.000 ± 0%      9.000 ± 0%        ~ (p=1.000 n=7) ¹
Shaping/100runes-latin-0fontCache-8                 9.000 ± 0%      9.000 ± 0%        ~ (p=1.000 n=7) ¹
Shaping/1000runes-latin-0fontCache-8                9.000 ± 0%      9.000 ± 0%        ~ (p=1.000 n=7) ¹
ShapingFontLoad/arabic:amiri_regular-8             69.11k ± 0%     43.76k ± 0%  -36.68% (p=0.001 n=7)
ShapingFontLoad/lating:go_regular-8                2.764k ± 0%     2.764k ± 0%        ~ (p=1.000 n=7) ¹
ShapingFontLoad/lating:roboto_regular-8           13.504k ± 0%     4.476k ± 0%  -66.85% (p=0.001 n=7)
ShapingGlyphInfoExtraction/10runes-arabic-8         163.0 ± 0%      163.0 ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/100runes-arabic-8       1.625k ± 0%     1.625k ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/1000runes-arabic-8      16.81k ± 0%     16.81k ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/10runes-latin-8          105.0 ± 0%      105.0 ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/100runes-latin-8        1.028k ± 0%     1.028k ± 0%        ~ (p=1.000 n=7) ¹
ShapingGlyphInfoExtraction/1000runes-latin-8       10.25k ± 0%     10.25k ± 0%        ~ (p=1.000 n=7) ¹
geomean                                                        ²                 -3.34%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

To sum up, #60 has the following impact relative to what we do now:

- Speed:
    - Shaping performs nearly identically, with perhaps a 1% speed penalty
    - 27% speedup in parsing amiri, our sample arabic font
    - 56% speedup in parsing roboto, our sample latin font
    - Glyph data extraction takes a 5% penalty for arabic and an 8% penalty for latin (backwards of what I'd expect)
- Memory use:
    - No significant impact on shaping memory use.
    - 25% memory use reduction for our sample arabic font.
    - 73% memory use reduction for roboto (sample latin font).
    - Interestingly, no savings for the go regular font.
    - Glyph info extraction does not appear to use more memory than it did before.
- Allocations:
    - No impact on shaping allocations.
    - 36% reduction in allocations for sample arabic font.
    - 66% reduction in allocations for roboto (sample latin font).
    - Again, no reduction for the Go font.
    - No change to glyph info extraction allocations

So on the whole, #60  reduces the footprint of some fonts significantly while imposing a slight 5-8% penalty on glyph info extraction. I think that sounds like a much better tradeoff. What do you all think?